Isolate filter feature for clean rollback; flag lex-comparison pitfall
Move all CQL filter and chunking logic out of api.py / utils.py into a
dedicated dataretrieval/waterdata/filters.py module (with chunked as a
decorator on the per-request fetch), and extract get_nearest_continuous
into a sibling nearest.py — so the entire filter feature can be removed
by deleting two source files, two test files, and two re-export lines.
Adds a pre-flight check that raises on unquoted-numeric comparisons
(value > 1000, parameter_code IN (60, 61), value BETWEEN 5 AND 10),
since every Water Data API queryable is string-typed and the server
either returns HTTP 500 or silently produces lexicographically-sorted
wrong rows.
