+++ dataretrieval/waterdata/types.py
@@ -40,8 +40,6 @@ SERVICES = Literal[
     "results",
 ]
 
-FILTER_LANG = Literal["cql-text", "cql-json"]
-
 PROFILES = Literal[
     "actgroup",
     "actmetric",
