+++ dataretrieval/waterdata/utils.py
@@ -4,10 +4,8 @@ import json
 import logging
 import os
 import re
-from collections.abc import Iterator
 from datetime import datetime
 from typing import Any, get_args
-from urllib.parse import quote_plus
 
 import pandas as pd
 import requests
@@ -15,6 +13,7 @@ from zoneinfo import ZoneInfo
 
 from dataretrieval import __version__
 from dataretrieval.utils import BaseMetadata
+from dataretrieval.waterdata import filters
 from dataretrieval.waterdata.types import (
     PROFILE_LOOKUP,
     PROFILES,
@@ -224,165 +223,6 @@ def _format_api_dates(
                 return formatted
     else:
         raise ValueError("datetime_input should only include 1-2 values")
-
-
-# Conservative fallback budget (characters) for a single CQL ``filter``
-# query parameter, used when the caller invokes ``_chunk_cql_or`` without
-# a ``max_len``. ``get_ogc_data`` computes a tighter per-request budget
-# from ``_WATERDATA_URL_BYTE_LIMIT`` below.
-_CQL_FILTER_CHUNK_LEN = 5000
-
-# Total URL byte limit the Water Data API will accept before replying
-# HTTP 414 (Request-URI Too Large). Empirically the cliff sits at
-# ~8,200 bytes of full URL, which lines up with nginx's default
-# ``large_client_header_buffers`` of 8 KB (8192). 8000 leaves ~200 bytes
-# of headroom for request-line framing ("GET ... HTTP/1.1\r\n") and any
-# intermediate proxy variance.
-_WATERDATA_URL_BYTE_LIMIT = 8000
-
-# Conservative over-estimate of the URL bytes consumed by everything
-# *except* the filter value — the base URL, other query params, and the
-# ``&filter=`` / ``&filter-lang=...`` keys. Used only to decide whether a
-# filter is small enough that the expensive budget probe can be skipped.
-_NON_FILTER_URL_HEADROOM = 1000
-
-
-def _iter_or_boundaries(expr: str) -> Iterator[tuple[int, int]]:
-    """Yield ``(start, end)`` spans of each top-level ``OR`` separator.
-
-    Tracks single/double-quoted string literals and parenthesized
-    sub-expressions so that ``OR`` tokens inside them are skipped.
-    Matching is case-insensitive and the yielded span covers the
-    surrounding whitespace on both sides.
-    """
-    depth = 0
-    in_quote = None
-    i = 0
-    n = len(expr)
-    while i < n:
-        ch = expr[i]
-        if in_quote is not None:
-            if ch == in_quote:
-                in_quote = None
-            i += 1
-            continue
-        if ch in ("'", '"'):
-            in_quote = ch
-            i += 1
-            continue
-        if ch == "(":
-            depth += 1
-            i += 1
-            continue
-        if ch == ")":
-            depth -= 1
-            i += 1
-            continue
-        if depth == 0 and ch.isspace():
-            j = i + 1
-            while j < n and expr[j].isspace():
-                j += 1
-            if j + 2 <= n and expr[j : j + 2].lower() == "or":
-                k = j + 2
-                if k < n and expr[k].isspace():
-                    m = k + 1
-                    while m < n and expr[m].isspace():
-                        m += 1
-                    yield i, m
-                    i = m
-                    continue
-        i += 1
-
-
-def _split_top_level_or(expr: str) -> list[str]:
-    """Split a CQL expression at each top-level ``OR`` separator.
-
-    Respects parentheses and single/double-quoted string literals so that
-    ``OR`` tokens inside ``(A OR B)`` or ``'word OR word'`` are left alone.
-    Matching is case-insensitive. Whitespace around each emitted part is
-    stripped; empty parts are dropped.
-    """
-    parts = []
-    last = 0
-    for start, end in _iter_or_boundaries(expr):
-        parts.append(expr[last:start].strip())
-        last = end
-    parts.append(expr[last:].strip())
-    return [p for p in parts if p]
-
-
-def _chunk_cql_or(expr: str, max_len: int = _CQL_FILTER_CHUNK_LEN) -> list[str]:
-    """Split a CQL expression into OR-chunks that each fit under ``max_len``.
-
-    The splitter only understands top-level ``OR`` chains, since that is
-    the only shape that can be recombined losslessly as a disjunction of
-    independent sub-queries. Returns ``[expr]`` unchanged when the whole
-    expression already fits, when it contains no top-level ``OR``, or when
-    any single clause is larger than ``max_len`` on its own (we would
-    rather send a too-long request and surface the server's 414 than
-    silently drop data).
-    """
-    if len(expr) <= max_len:
-        return [expr]
-    parts = _split_top_level_or(expr)
-    if len(parts) < 2 or any(len(p) > max_len for p in parts):
-        return [expr]
-
-    chunks = []
-    current = []
-    current_len = 0
-    for part in parts:
-        join_cost = len(" OR ") if current else 0
-        if current and current_len + join_cost + len(part) > max_len:
-            chunks.append(" OR ".join(current))
-            current = [part]
-            current_len = len(part)
-        else:
-            current.append(part)
-            current_len += join_cost + len(part)
-    if current:
-        chunks.append(" OR ".join(current))
-    return chunks
-
-
-def _effective_filter_budget(args: dict[str, Any], filter_expr: str) -> int:
-    """Compute the raw CQL byte budget for ``filter_expr`` in this request.
-
-    The server limits total URL length (see ``_WATERDATA_URL_BYTE_LIMIT``),
-    not raw CQL length. To derive a raw-byte budget we can hand to
-    ``_chunk_cql_or``:
-
-    1. Probe the URL space consumed by the other query params by building
-       the request with a 1-byte placeholder filter.
-    2. Subtract from the URL limit to get the bytes available for the
-       encoded filter value.
-    3. Convert back to raw CQL bytes using the *maximum* per-clause
-       encoding ratio, not the whole-filter average. A chunk can end up
-       containing only the heavier-encoding clauses (e.g. heavy ones
-       clustered at one end of the filter), so budgeting against the
-       average lets such a chunk overflow the URL limit by a few bytes.
-    """
-    # Fast path: if the whole encoded filter already fits with room for
-    # any plausible non-filter URL overhead, skip the probe and the
-    # splitter entirely. Signals pass-through via a budget larger than
-    # the filter. Saves a PreparedRequest build + a full splitter scan
-    # on every short-filter call.
-    encoded_len = len(quote_plus(filter_expr))
-    if encoded_len + _NON_FILTER_URL_HEADROOM <= _WATERDATA_URL_BYTE_LIMIT:
-        return len(filter_expr) + 1
-
-    probe = _construct_api_requests(**{**args, "filter": "x"})
-    non_filter_url_bytes = len(probe.url) - 1
-    available_url_bytes = _WATERDATA_URL_BYTE_LIMIT - non_filter_url_bytes
-    if available_url_bytes <= 0:
-        # The non-filter URL already exceeds the byte limit, so no chunk
-        # we could produce would fit. Return a budget larger than the
-        # filter so _chunk_cql_or passes it through unchanged — one 414
-        # from the server is clearer than a burst of N failing sub-requests.
-        return len(filter_expr) + 1
-    parts = _split_top_level_or(filter_expr) or [filter_expr]
-    encoding_ratio = max(len(quote_plus(p)) / len(p) for p in parts if p)
-    return max(100, int(available_url_bytes / encoding_ratio))
 
 
 def _cql2_param(args: dict[str, Any]) -> str:
@@ -995,96 +835,30 @@ def get_ogc_data(
     convert_type = args.pop("convert_type", False)
     args = {k: v for k, v in args.items() if v is not None}
 
-    chunks = _plan_filter_chunks(args)
-    frames, responses = _fetch_chunks(args, chunks)
-
-    return_list = _combine_chunk_frames(frames)
+    return_list, response = _fetch_once(args)
     return_list = _deal_with_empty(return_list, properties, service)
     if convert_type:
         return_list = _type_cols(return_list)
     return_list = _arrange_cols(return_list, properties, output_id)
     return_list = _sort_rows(return_list)
 
-    return return_list, _aggregate_response_metadata(responses)
+    return return_list, BaseMetadata(response)
 
 
-def _plan_filter_chunks(args: dict[str, Any]) -> list[str | None]:
-    """Decide how to fan ``args["filter"]`` out across HTTP calls.
+@filters.chunked(build_request=_construct_api_requests)
+def _fetch_once(
+    args: dict[str, Any],
+) -> tuple[pd.DataFrame, requests.Response]:
+    """Send one prepared-args OGC request; return the frame + response.
 
-    Returns one entry per request to send. A ``None`` entry means "send
-    ``args`` as-is" — either there is no filter, or the filter language
-    is not one we can safely split (only cql-text top-level ``OR``
-    chains are chunkable). Otherwise each string entry is a chunked
-    cql-text expression that replaces ``args["filter"]`` for its
-    sub-request. Overlapping user OR-clauses are deduplicated by feature
-    id later in ``_combine_chunk_frames``.
+    Filter chunking is added orthogonally by the ``@filters.chunked``
+    decorator: with no filter (or an un-chunkable one) the decorator
+    passes ``args`` through to this body; with a chunkable filter it
+    fans out and calls this body once per sub-filter, then combines.
+    Either way the return shape is ``(frame, response)``.
     """
-    filter_expr = args.get("filter")
-    filter_lang = args.get("filter_lang")
-    chunkable = (
-        isinstance(filter_expr, str)
-        and filter_expr
-        and filter_lang in {None, "cql-text"}
-    )
-    if not chunkable:
-        return [None]
-    raw_budget = _effective_filter_budget(args, filter_expr)
-    return _chunk_cql_or(filter_expr, max_len=raw_budget)
-
-
-def _fetch_chunks(
-    args: dict[str, Any], chunks: list[str | None]
-) -> tuple[list[pd.DataFrame], list[requests.Response]]:
-    """Send one request per chunk; return the per-chunk frames and responses."""
-    frames: list[pd.DataFrame] = []
-    responses: list[requests.Response] = []
-    for chunk in chunks:
-        chunk_args = args if chunk is None else {**args, "filter": chunk}
-        req = _construct_api_requests(**chunk_args)
-        frame, response = _walk_pages(geopd=GEOPANDAS, req=req)
-        frames.append(frame)
-        responses.append(response)
-    return frames, responses
-
-
-def _combine_chunk_frames(frames: list[pd.DataFrame]) -> pd.DataFrame:
-    """Concatenate per-chunk frames, handling the edge cases.
-
-    Drops empty frames before concat — ``_get_resp_data`` returns a
-    plain ``pd.DataFrame()`` on empty responses, which would downgrade
-    a concat of real GeoDataFrames back to a plain DataFrame and strip
-    geometry/CRS. Also dedups on the pre-rename feature ``id`` so
-    overlapping user-supplied OR-clauses don't produce duplicate rows
-    across chunks.
-    """
-    non_empty = [f for f in frames if not f.empty]
-    if not non_empty:
-        return pd.DataFrame()
-    if len(non_empty) == 1:
-        return non_empty[0]
-    combined = pd.concat(non_empty, ignore_index=True)
-    if "id" in combined.columns:
-        combined = combined.drop_duplicates(subset="id", ignore_index=True)
-    return combined
-
-
-def _aggregate_response_metadata(
-    responses: list[requests.Response],
-) -> BaseMetadata:
-    """Build metadata from the first response, summing elapsed across chunks.
-
-    The first response's URL and headers are the representative ones to
-    return. When the filter was fanned across multiple chunks, replace
-    its elapsed with the sum so ``query_time`` reflects the full
-    operation rather than just the first sub-request.
-    """
-    metadata_response = responses[0]
-    if len(responses) > 1:
-        metadata_response.elapsed = sum(
-            (r.elapsed for r in responses[1:]),
-            start=metadata_response.elapsed,
-        )
-    return BaseMetadata(metadata_response)
+    req = _construct_api_requests(**args)
+    return _walk_pages(geopd=GEOPANDAS, req=req)
 
 
 def _handle_stats_nesting(
